requirements.txt
django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.5.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.7.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
Pillow==10.4.0